requirements.txt
requests==2.25.1
python-dotenv==0.17.1
structlog==21.1.0
